fix: release-please's tagPullRequestNumber (#1155)
With this `tagPullRequestNumber: true` configuration, the Release Please bot will tag the commit SHA with "release-please-<PR number>`.

This information is useful when we create an automation to report the release status back to the pull request.

This configuration change is not exactly a bug fix but "fix:" prefix gives a new release.